Prime number finder matrix vector.
*******************************************************
****** About Formula ********************************
*******************************************************
*****  The above formula, the prime numbers in numbers plane
*****  Ax, Ay or Az. Results obtained from the above formula,
*****  which will be advantageous to use the unit automatically
*****  shows the analytical plane. Vector in the first series
*****  of the formula, primes numbers say bid welcome to us.
*****  
****** 3L+4L tangential
****** 4L+3L tangential
****** 3L+2L+3L semicircle
****** 3L+4L+3L+4L rectangle
****** 5L + in center 4L perpendicular
********************************************************

Prime numbers that can be called to create a unique geometric 
shapes and planes. Therefore, it is difficult to find
the prime numbers with a simple method. As consisting of 
prime numbers, numbers, geometric shapes, and the plane can be
divided into only their coats. Therefore, you should look at 
a number can create unique geometric shapes to create a unique
checking the inertness. Primes are special, try to understand it 
helps to understand the math, and even provides even of
mathematical development. For example, to invent systems that 
can be found prime numbers of its number in numeretical system.

Important topics in this so primality:
* Lack of self divisor other numbers.
* Unique dimentional plane that can be created.
* Unique geometric shape that can be created.
* Angular, linear, area measurement accuracy.

Sample:
34 numbers divisible only 17 and 2, and unique heptadecagon, 
17 is no integer divisor in lower group. 38 numbers
divisible only 19 and 2, and unique enneadecagon, 19 is no integer 
divisor in lower group.